rtic-auth-server/requirements.txt
gunicorn==22.0.0
firebase-admin==6.5.0
fastapi==0.115.0
uvicorn==0.30.6
Werkzeug==3.0.3
google-cloud-firestore==2.17.2
firebase-admin==6.5.0
python-dotenv==1.0.1
pydantic[email]==2.9.2
passlib==1.7.4